refactored metabolomics.py
