+++ src/nplinker/metabolomics/metabolomics.py
@@ -15,277 +15,22 @@
 import csv
 import os
 import re
-from .annotations import GNPS_DATA_COLUMNS
-from .annotations import GNPS_KEY
-from .annotations import create_gnps_annotation
-from .logconfig import LogConfig
-from .parsers.mgf import LoadMGF
-from .strains import StrainCollection
-from .utils import sqrt_normalise
+from nplinker.annotations import GNPS_DATA_COLUMNS
+from nplinker.annotations import GNPS_KEY
+from nplinker.annotations import create_gnps_annotation
+from nplinker.logconfig import LogConfig
+from nplinker.parsers.mgf import LoadMGF
+
+from .Spectrum import Spectrum
+from .MolecularFamily import MolecularFamily
+from .SingletonFamily import SingletonFamily
+
 
 
 logger = LogConfig.getLogger(__file__)
 
-JCAMP = '##TITLE={}\\n' +\
-        '##JCAMP-DX=nplinker vTODO\\n' +\
-        '##DATA TYPE=Spectrum\\n' +\
-        '##DATA CLASS=PEAKTABLE\\n' +\
-        '##ORIGIN=TODO_DATASET_ID\\n' +\
-        '##OWNER=nobody\\n' +\
-        '##XUNITS=M/Z\\n' +\
-        '##YUNITS=RELATIVE ABUNDANCE\\n' +\
-        '##NPOINTS={}\\n' +\
-        '##PEAK TABLE=(XY..XY)\\n' +\
-        '{}\\n' +\
-        '##END=\\n'
-
 # compile a regex for matching .mzXML and .mzML strings
 RE_MZML_MZXML = re.compile('.mzXML|.mzML')
-
-
-class Spectrum():
-
-    def __init__(self,
-                 id,
-                 peaks,
-                 spectrum_id,
-                 precursor_mz,
-                 parent_mz=None,
-                 rt=None):
-        self.id = id
-        self.peaks = sorted(peaks, key=lambda x: x[0])  # ensure sorted by mz
-        self.normalised_peaks = sqrt_normalise(self.peaks)  # useful later
-        self.n_peaks = len(self.peaks)
-        self.max_ms2_intensity = max(
-            intensity for mz, intensity in self.peaks)
-        self.total_ms2_intensity = sum(
-            intensity for mz, intensity in self.peaks)
-        assert (isinstance(spectrum_id, int))
-        self.spectrum_id = spectrum_id
-        self.rt = rt
-        self.precursor_mz = precursor_mz
-        self.parent_mz = parent_mz
-        self.gnps_id = None  # CCMSLIB...
-        # TODO should add intensity here too
-        self.metadata = {}
-        self.edges = []
-        self.strains = StrainCollection()
-        # this is a dict indexed by Strain objects (the strains found in this Spectrum), with
-        # the values being dicts of the form {growth_medium: peak intensity} for the parent strain
-        self.growth_media = {}
-        self.family_id = -1
-        self.family = None
-        # a dict indexed by filename, or "gnps"
-        self.annotations = {}
-        self._losses = None
-        self._jcamp = None
-
-    def add_strain(self, strain, growth_medium, peak_intensity):
-        # adds the strain to the StrainCollection if not already there
-        self.strains.add(strain)
-
-        if strain not in self.growth_media:
-            self.growth_media[strain] = {}
-
-        if growth_medium is None:
-            self.growth_media[strain].update({
-                f'unknown_medium_{len(self.growth_media[strain])}':
-                peak_intensity
-            })
-            return
-
-        if strain in self.growth_media and growth_medium in self.growth_media[
-                strain]:
-            raise Exception('Growth medium clash: {} / {} {}'.format(
-                self, strain, growth_medium))
-
-        self.growth_media[strain].update({growth_medium: peak_intensity})
-
-    @property
-    def is_library(self):
-        return GNPS_KEY in self.annotations
-
-    def set_annotations(self, key, data):
-        self.annotations[key] = data
-
-    @property
-    def gnps_annotations(self):
-        if GNPS_KEY not in self.annotations:
-            return None
-
-        return self.annotations[GNPS_KEY][0]
-
-    def has_annotations(self):
-        return len(self.annotations) > 0
-
-    def get_metadata_value(self, key):
-        val = self.metadata.get(key, None)
-        return val
-
-    def has_strain(self, strain):
-        return strain in self.strains
-
-    def get_growth_medium(self, strain):
-        if strain not in self.strains:
-            return None
-
-        gms = self.growth_media[strain]
-        return list(gms.keys())[0]
-
-    def to_jcamp_str(self, force_refresh=False):
-        if self._jcamp is not None and not force_refresh:
-            return self._jcamp
-
-        peakdata = '\\n'.join('{}, {}'.format(*p) for p in self.peaks)
-        self._jcamp = JCAMP.format(str(self), self.n_peaks, peakdata)
-        return self._jcamp
-
-    def __str__(self):
-        return "Spectrum(id={}, spectrum_id={}, strains={})".format(
-            self.id, self.spectrum_id, len(self.strains))
-
-    def __repr__(self):
-        return str(self)
-
-    def __eq__(self, other):
-        return self.id == other.id
-
-    def __hash__(self):
-        return self.id
-
-    def __cmp__(self, other):
-        if self.parent_mz >= other.parent_mz:
-            return 1
-        else:
-            return -1
-
-    def __lt__(self, other):
-        if self.parent_mz <= other.parent_mz:
-            return 1
-        else:
-            return 0
-
-    # from molnet repo
-    def keep_top_k(self, k=6, mz_range=50):
-        # only keep peaks that are in the top k in += mz_range
-        start_pos = 0
-        new_peaks = []
-        for mz, intensity in self.peaks:
-            while self.peaks[start_pos][0] < mz - mz_range:
-                start_pos += 1
-            end_pos = start_pos
-
-            n_bigger = 0
-            while end_pos < len(
-                    self.peaks) and self.peaks[end_pos][0] <= mz + mz_range:
-                if self.peaks[end_pos][1] > intensity:
-                    n_bigger += 1
-                end_pos += 1
-
-            if n_bigger < k:
-                new_peaks.append((mz, intensity))
-
-        self.peaks = new_peaks
-        self.n_peaks = len(self.peaks)
-        if self.n_peaks > 0:
-            self.normalised_peaks = sqrt_normalise(self.peaks)
-            self.max_ms2_intensity = max(
-                intensity for mz, intensity in self.peaks)
-            self.total_ms2_intensity = sum(
-                intensity for mz, intensity in self.peaks)
-        else:
-            self.normalised_peaks = []
-            self.max_ms2_intensity = 0.0
-            self.total_ms2_intensity = 0.0
-
-    @property
-    def losses(self):
-        """
-        All mass shifts in the spectrum, and the indices of the peaks
-        """
-        if self._losses is None:
-            # populate loss table
-            losses = []
-            for i in range(len(self.peaks)):
-                loss = self.precursor_mz - self.peaks[i][0]
-                losses.append((loss, self.id, i))
-
-
-# THIS SEEMED TO ME LIKE IT WOULD TAKE THE WRONG DIFFERENCES AS LOSSES:
-# TODO: please check!
-#                for j in range(i):
-#                    loss = self.peaks[i][0] - self.peaks[j][0]
-#                    losses.append((loss, i, j))
-
-# Sort by loss
-            losses.sort(key=lambda x: x[0])
-            self._losses = losses
-        return self._losses
-
-    def has_loss(self, mass, tol):
-        """
-        Check if the scan has the specified loss (within tolerance)
-        """
-        matched_losses = []
-
-        idx = 0
-        # Check losses in range [0, mass]
-        while idx < len(self.losses) and self.losses[idx][0] <= mass:
-            if mass - self.losses[idx][0] < tol:
-                matched_losses.append(self.losses[idx])
-            idx += 1
-
-        # Add all losses in range [mass, mass+tol(
-        while idx < len(self.losses) and self.losses[idx][0] < mass + tol:
-            matched_losses.append(self.losses[idx])
-            idx += 1
-
-        return matched_losses
-
-
-class MolecularFamily():
-
-    def __init__(self, family_id):
-        self.id = -1
-        self.family_id = family_id
-        self.spectra = []
-        self.family = None
-
-    def has_strain(self, strain):
-        for spectrum in self.spectra:
-            if spectrum.has_strain(strain):
-                return True
-
-        return False
-
-    @property
-    def strains(self):
-        strains = set()
-        for spectrum in self.spectra:
-            strains = strains.union(spectrum.strains)
-        return strains
-
-    def add_spectrum(self, spectrum):
-        self.spectra.append(spectrum)
-
-    def __str__(self):
-        return 'MolFam(family_id={}, spectra={})'.format(
-            self.family_id, len(self.spectra))
-
-    def __eq__(self, other):
-        return self.id == other.id
-
-    def __hash__(self):
-        return self.id
-
-
-class SingletonFamily(MolecularFamily):
-
-    def __init__(self):
-        super().__init__(-1)
-
-    def __str__(self):
-        return f"Singleton molecular family (id={self.id})"
 
 
 #
@@ -298,7 +43,7 @@ GNPS_FORMAT_OLD_UNIQUEFILES = 'uniquefiles'
 GNPS_FORMAT_NEW_FBMN = 'fbmn'
 
 
-def get_headers(filename, delimiters=['\t', ',']):
+def _get_headers(filename, delimiters=['\t', ',']):
     headers = None
     with open(filename) as f:
         headers = f.readline()
@@ -310,8 +55,8 @@ def get_headers(filename, delimiters=['\t', ',']):
     return headers
 
 
-def identify_gnps_format(filename, has_quant_table):
-    headers = get_headers(filename)
+def _identify_gnps_format(filename, has_quant_table):
+    headers = _get_headers(filename)
 
     if headers is None:
         return GNPS_FORMAT_UNKNOWN
@@ -402,7 +147,7 @@ def load_edges(edges_file, spec_dict):
                 spec1.family_id = family
 
 
-def messy_strain_naming_lookup(mzxml, strains):
+def _messy_strain_naming_lookup(mzxml, strains):
     if mzxml in strains:
         # life is simple!
         return strains.lookup(mzxml)
@@ -435,7 +180,7 @@ def messy_strain_naming_lookup(mzxml, strains):
     return None
 
 
-def md_convert(val):
+def _md_convert(val):
     """Try to convert raw metadata values from text to integer, then float if that fails"""
     try:
         return int(val)
@@ -448,7 +193,7 @@ def md_convert(val):
     return val
 
 
-def parse_mzxml_header(hdr, strains, md_table, ext_metadata_parsing):
+def _parse_mzxml_header(hdr, strains, md_table, ext_metadata_parsing):
     # given a column header from the quantification_table file produced by GNPS,
     # this method checks if it's one that contains peak information for a given
     # strain label by searching for the 'Peak area' string.
@@ -513,7 +258,7 @@ def parse_mzxml_header(hdr, strains, md_table, ext_metadata_parsing):
     return (strain_name, growth_medium, strain_name not in strains)
 
 
-def load_clusterinfo_old(gnps_format, strains, filename, spec_dict):
+def _load_clusterinfo_old(gnps_format, strains, filename, spec_dict):
     # each line of this file represents a metabolite.
     # columns representing strain IDs are *ignored* here in favour of parsing
     # .mz(X)ML filenames from either the AllFiles or UniqueFileSources column.
@@ -556,7 +301,7 @@ def load_clusterinfo_old(gnps_format, strains, filename, spec_dict):
                 metadata['files'][mzxml] = mzxml
 
                 # should have a strain alias for the mxXML file to lookup here (in theory at least)
-                strain = messy_strain_naming_lookup(mzxml, strains)
+                strain = _messy_strain_naming_lookup(mzxml, strains)
                 if strain is None:
                     # TODO: how to handle this? can't just give up, simply warn?
                     if mzxml not in unknown_strains:
@@ -584,7 +329,7 @@ def load_clusterinfo_old(gnps_format, strains, filename, spec_dict):
     return unknown_strains
 
 
-def parse_metadata_table(filename):
+def _parse_metadata_table(filename):
     """Parses the metadata table file from GNPS"""
     if filename is None:
         return None
@@ -647,12 +392,12 @@ def parse_metadata_table(filename):
     return table
 
 
-def load_clusterinfo_fbmn(strains, nodes_file, extra_nodes_file, md_table_file,
+def _load_clusterinfo_fbmn(strains, nodes_file, extra_nodes_file, md_table_file,
                           spec_dict, ext_metadata_parsing):
     spec_info = {}
 
     # parse metadata table if available
-    md_table = parse_metadata_table(md_table_file)
+    md_table = _parse_metadata_table(md_table_file)
 
     # combine the information in the nodes_file (clusterinfo_summary folder) and
     # the extra_nodes_file (quantification_table folder), indexed by the "cluster index"
@@ -728,7 +473,7 @@ def load_clusterinfo_fbmn(strains, nodes_file, extra_nodes_file, md_table_file,
             # matching Strain object from the set of mappings we have (optionally doing
             # some further stuff involving the metadata table)
             (strain_name, growth_medium,
-             is_unknown) = parse_mzxml_header(k, strains, md_table,
+             is_unknown) = _parse_mzxml_header(k, strains, md_table,
                                               ext_metadata_parsing)
 
             # if the strain name couldn't be recognised at all, ignore it
@@ -741,7 +486,7 @@ def load_clusterinfo_fbmn(strains, nodes_file, extra_nodes_file, md_table_file,
                 continue
 
             # create a new strain object if the intensity value is a float > 0
-            v = md_convert(v)
+            v = _md_convert(v)
             if strain_name in strains and isinstance(v, float) and v > 0:
                 # find the strain object, and add the growth medium + intensity to it. the
                 # growth medium will only be set if extended_metadata_table_parsing is
@@ -777,7 +522,7 @@ def load_dataset(strains,
     # add edges info to the spectra
     load_edges(edges_file, spec_dict)
 
-    gnps_format = identify_gnps_format(nodes_file, quant_table_file
+    gnps_format = _identify_gnps_format(nodes_file, quant_table_file
                                        is not None)
     logger.debug('Nodes_file: {}, quant_table_exists?: {}'.format(
         nodes_file, quant_table_file is None))
@@ -788,21 +533,21 @@ def load_dataset(strains,
     # if we don't have a quantification table, must be older-style dataset (or unknown format)
     if gnps_format != GNPS_FORMAT_NEW_FBMN and quant_table_file is None:
         logger.info('Found older-style GNPS dataset, no quantification table')
-        unknown_strains = load_clusterinfo_old(gnps_format, strains,
+        unknown_strains = _load_clusterinfo_old(gnps_format, strains,
                                                nodes_file, spec_dict)
     else:
         logger.info('quantification table exists, new-style GNPS dataset')
-        _, unknown_strains = load_clusterinfo_fbmn(strains, nodes_file,
+        _, unknown_strains = _load_clusterinfo_fbmn(strains, nodes_file,
                                                    quant_table_file,
                                                    metadata_table_file,
                                                    spec_dict,
                                                    ext_metadata_parsing)
 
-    molfams = make_families(spectra)
+    molfams = _make_families(spectra)
     return spec_dict, spectra, molfams, unknown_strains
 
 
-def make_families(spectra):
+def _make_families(spectra):
     families = []
     family_dict = {}
     family_index = 0
